+++ src/components/reusable/table/index.css
@@ -1,8 +1,8 @@
-.ant-table-row:nth-child(even)
+.ant-table-row:nth-child(odd)
 {
   background-color: #EEEDFF;
  }
- .ant-table-row:nth-child(add)
+ .ant-table-row:nth-child(even)
 {
   background-color: #fff;
  }
