+++ src/pages/setStandards/index.css
@@ -1,0 +1,35 @@
+.operatingRange {
+  word-spacing: 1rem;
+}
+
+.thresholdValue {
+  word-spacing: 1rem;
+}
+
+.ant-table-thead > tr > th:nth-child(6) {
+  word-spacing: normal;
+}
+.ant-table-thead > tr > th:nth-child(5) {
+  word-spacing: normal;
+}
+
+.ant-table-thead {
+  font-weight: 400;
+  font-size: 16px;
+  font-family: 'Gotham-medium';
+}
+
+table {
+  margin-bottom: 17px;
+}
+
+.ant-pagination-item-active a {
+  background-color: #5f2eea;
+  color: white !important;
+}
+.ant-table-cell {
+  color: #444444;
+  font-weight: 400;
+  font-size: 16px;
+  font-family: 'Gotham-light';
+}
